python/serving/serving_framework/requirements.txt
absl-py~=2.1.0
flask~=3.0.3
grpcio~=1.67.1
gunicorn~=23.0.0
numpy~=2.0.2
requests~=2.32.3
# TODO: b/375469331 - Enable testing with most current requests-mock release.
requests-mock==1.9.3
tensorflow~=2.18.0
tensorflow-serving-api~=2.17.0
typing-extensions~=4.12.2
